refactored task_api.py, moved classes into game_state.py predicate.py
