+++ nmmo/task/__init__.py
@@ -1,3 +1,5 @@
+from .game_state import *
+from .predicate import *
 from .task_api import *
 from .sampler import *
 from .base_predicate import *
